Moved plots to separate file
